// src/infra/http/protocols/http-client.types.ts
import { AxiosResponse } from "axios";

export type HttpRequest = {
    url: string;
    method: HttpMethod;
    body?: any;
    headers?: any;
    params?: any;
};

export type HttpMethod = 'post' | 'get' | 'put' | 'delete';

export enum HttpStatusCode {
    ok = 200,
    noContent = 204,
    badRequest = 400,
    unauthorized = 401,
    forbidden = 403,
    notFound = 404,
    serverError = 500
}

export interface HttpClient<R = any> {
    request: (data: HttpRequest) => Promise<AxiosResponse<R>>;
}